+++ src/add_or_insert/add_or_insert.rs
@@ -1,0 +1,5 @@
+
+pub trait AddOrInsert<K, V> {
+    fn add_or_insert(&mut self, key: K, value:V);
+    
+}