+++ src/get_many_mut/get_many_mut.rs
@@ -1,6 +1,3 @@
-mod btreemap;
-mod vec;
-
 pub trait GetManyMut<T, Q, const N: usize>{
     fn get_many_mut(&mut self, keys: [Q; N]) -> Option<[&mut T; N]>;  
 }