+++ src/get_many_mut/mod.rs
@@ -1,0 +1,4 @@
+mod get_many_mut;
+pub use get_many_mut::GetManyMut;
+mod btreemap;
+mod vec;