+++ src/push_or_insert/hashmap.rs
@@ -1,9 +1,7 @@
 use std::{collections::HashMap, hash::Hash};
 
-pub trait PushOrInsert<K: Hash + Eq, V>{
-    fn append_or_insert(&mut self, k: K, v: &mut Vec<V>);
-    fn push_or_insert(&mut self, k: K, v: V);
-}
+use super::PushOrInsert;
+
 impl<K: Hash + Eq, V> PushOrInsert<K, V> for HashMap<K, Vec<V>> {
     fn append_or_insert(&mut self, k: K, v: &mut Vec<V>){
         let Some(thing)= self.get_mut(&k) else {
@@ -25,5 +23,28 @@ impl<K: Hash + Eq, V> PushOrInsert<K, V> for HashMap<K, Vec<V>> {
             return;
         };
         thing.push(v);
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    #[test]
+    fn add(){
+        let mut specimen = HashMap::from([("a" ,vec![1])]);
+        specimen.push_or_insert("a", 2);
+        assert_eq!(
+            specimen,
+            HashMap::from([("a", vec![1,2])])
+        )
+    }
+    #[test]
+    fn insert(){
+        let mut specimen = HashMap::new();
+        specimen.push_or_insert("a", 3);
+        assert_eq!(
+            specimen,
+            HashMap::from([("a", vec![3])])
+        )
     }
 }